Roles permissions

+++ app/Http/Controllers/Seller/SellerVariantController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Seller;
 
-use App\Models\User;
+
 use App\Models\Product;
 use App\Models\Variant;
 use App\Services\UploadImageService;
@@ -10,6 +10,7 @@ use App\Http\Controllers\ApiController;
 use App\Http\Requests\StoreVariantRequest;
 use App\Http\Requests\UpdateVariantRequest;
 use App\Http\Resources\VariantResource;
+use App\Models\Attribute;
 
 class SellerVariantController extends ApiController
 {
@@ -33,23 +34,19 @@ class SellerVariantController extends ApiController
      * @return \Illuminate\Http\Response
      */
     public function store(StoreVariantRequest $request, Product $product){
-        //validate variant details and image details
-        $variant_data = $request->validated();
 
-        $images = $request->file('images');
+        $request->validated();
+        $variant_data = $request->except('attrs','medias');
 
+        $images = $request->file('medias');
 
         $variant_data['product_id'] = $product->id;
 
-        
         $newVariant = Variant::create($variant_data);
 
-        // $attr = array_map('intval', explode(',', $request->attrs));
-        $attr = $request->attrs;
+        $attrs = Attribute::all()->whereIn('uuid', $request->attrs)->pluck('id');
+        $newVariant->attributes()->sync($attrs);
 
-        $newVariant->attributes()->sync($attr);
-
-        //send images to be uploaded
         UploadImageService::upload($newVariant,$images, Variant::class);
 
         return $this->showOne(new VariantResource($newVariant));
@@ -62,28 +59,30 @@ class SellerVariantController extends ApiController
      * @param  Variant $variant
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateVariantRequest $request,Variant $variant){
-   
+    public function update(UpdateVariantRequest $request,Variant $variant)
+    {
+        $this->authorize('update', $variant);
         $request->validated();
-        $seller_id = $request->seller_id;
-        $images = null;
-        $images = $request->images;
-    
-        if($request->has('images')){
-            $images = $request->images;
-            $request_images = count($request->file('images'));
-            abort_if( $request_images > 1, 422, 'Can not have more than 1 image per variant');
+        $images = $request->medias;
+        if($request->has('medias')){
+            $images = $request->medias;
+            $request_images = count($request->file('medias'));
+            abort_if( $request_images > 1, 422, 'Can not update more than 1 image per variant');
             UploadImageService::upload($variant,$images, Variant::class);
         }
         
         
-        $variant->fill($request->except(['categories']));   
-        //get string ids and convert them to integer
-        $this->checkSeller($seller_id,$variant->product_id);
+        $variant->fill($request->except(['categories', 'attrs', 'medias', 'product_id']));   
 
-        $attr = $request->attrs;
+        if($request->has('product_id')){
+            $product = Product::where('uuid', $request->product_id)->firstOrFail();
+            $variant->product_id = $product->id;
+        }
 
-        $variant->attributes()->sync($attr);
+
+        $attrs = Attribute::all()->whereIn('uuid', $request->attrs)->pluck('id');
+        
+        $variant->attributes()->sync($attrs);
 
         $variant->save();
 
@@ -92,28 +91,16 @@ class SellerVariantController extends ApiController
     }
 
     /**
-     * @param Product $product
      * @param Variant $variant
      * 
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Product $product, Variant $variant)
+    public function destroy(Variant $variant)
     {   
-      
-        //ADD seller id of logged user or check if ADMIN role 
-        $seller_id = 4;
-        $this->checkSeller($seller_id, $product->id);
+        $this->authorize('delete', $variant);
 
         $variant->delete();
 
         return $this->showOne(new VariantResource($variant));
-    }
-
-    protected function checkSeller($seller_id, $product_id){
-        $product = Product::findOrFail($product_id);
-        $seller = User::findOrFail($seller_id);
-        abort_if($seller->id != $product->seller_id,
-        422,
-        'The specified seller is not the seller of this product!');
     }
 }
